// projects/common/src/lib/enums/utils/hermes.enum.ts
export enum HermesEnum {
    LOGIN = 'user-login',
    LOGOUT = 'user-logout',
    LOGOUT_AND_REDIRECT = 'user-logout-and-redirect',
    SELECT_ROLE = 'user-select-role',
    CHANGE_ROLE = 'user-change-role',
    CHANGE_PASSWORD = 'user-change-password',
    UPDATE_CURRENT_ROLE = 'user-update-role',
    UPDATE_ROLES = 'user-update-roles',
    COMPANY_ENABLE_DISABLE = 'company-enable-disable',

    CUSTOMER_LOGIN = 'customer-login',
    CUSTOMER_CHANGE_COMPANY = 'customer-change-company',
    CUSTOMER_LOGOUT = 'customer-logout',
    CUSTOMER_LOGOUT_AND_REDIRECT = 'customer-logout-and-redirect',
    CUSTOMER_UPDATE_PERMISSIONS = 'customer-update-permissions',

    CUSTOMER_CHANGE_THEME = 'customer-change-theme',
}